Update README.md byte array/slice wording (#192)
In go dynamic arrays are called slices which are really a pointer to a a {length, capacity, static array}. Regular arrays are a static size and cannot grow or shrink. I looked at the code and it does in fact seem to be using slices for entries. It was confusing to see 

See the go blog for further info https://blog.golang.org/go-slices-usage-and-internals