Fix atomic usage on 32-bit systems
The atomic package requires aligned variables (as per the package
documentation), but this is not happening for the variables on the
stack. Instead change the types for these to 32-bit variables as
there should be no need to track 64-bit stats.

This fixes various panics from the unaligned memory accesses.